requirements.txt
accelerate
datasets
torchmetrics
evaluate
transformers
wandb
pip
peft
nodelist-inflator
deepspeed
sentencepiece
jsonlines
discord.py
openai